Split self containing server into api and server modules
